coinbase-producer/requirements.txt
websocket-client==1.6.4
confluent-kafka==2.3.0